+++ DemoApp/Screens/UserProfileViewController.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2023 Stream.io Inc. All rights reserved.
+// Copyright © 2024 Stream.io Inc. All rights reserved.
 //
 
 import StreamChat
@@ -40,7 +40,7 @@ class UserProfileViewController: UIViewController, CurrentChatUserControllerDele
         updateButton.backgroundColor = .systemBlue
         updateButton.contentEdgeInsets = UIEdgeInsets(top: 0.0, left: 15, bottom: 0.0, right: 15)
         updateButton.addTarget(self, action: #selector(didTapUpdateButton), for: .touchUpInside)
-        updateButton.isHidden = !DemoAppConfiguration.isStreamInternalConfiguration
+        updateButton.isHidden = !StreamRuntimeCheck.isStreamInternalConfiguration
 
         NSLayoutConstraint.activate([
             imageView.widthAnchor.constraint(equalToConstant: 60),