+++ Sources/StreamChat/Database/DTOs/MessageDTO.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2023 Stream.io Inc. All rights reserved.
+// Copyright © 2024 Stream.io Inc. All rights reserved.
 //
 
 import CoreData
@@ -537,7 +537,7 @@ extension MessageDTO {
 
     var isLocalOnly: Bool {
         if let localMessageState = self.localMessageState {
-            return localMessageState.isWaitingToBeSentToServer
+            return localMessageState.isLocalOnly
         }
 
         return type == MessageType.ephemeral.rawValue || type == MessageType.error.rawValue