+++ Sources/StreamChatUI/ChatMessageList/Attachments/AttachmentViewCatalog.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2023 Stream.io Inc. All rights reserved.
+// Copyright © 2024 Stream.io Inc. All rights reserved.
 //
 
 import Foundation
@@ -15,8 +15,6 @@ open class AttachmentViewCatalog {
         message: ChatMessage,
         components: Components
     ) -> AttachmentViewInjector.Type? {
-        if message.isDeleted { return nil }
-
         let attachmentCounts = message.attachmentCounts
 
         if attachmentCounts.keys.contains(.image) || attachmentCounts.keys.contains(.video) {