+++ Tests/StreamChatTests/Workers/ChannelUpdater_Tests.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2023 Stream.io Inc. All rights reserved.
+// Copyright © 2024 Stream.io Inc. All rights reserved.
 //
 
 @testable import StreamChat
@@ -151,7 +151,7 @@ final class ChannelUpdater_Tests: XCTestCase {
         try database.createChannel(cid: cid, withMessages: false)
         // Local only message
         try database.createMessage(cid: cid, localState: .sendingFailed)
-        try database.createMessage(cid: cid, localState: .deletingFailed)
+        try database.createMessage(cid: cid, localState: .pendingSend)
         // Not local only message
         try database.createMessage(cid: cid, localState: .syncing)
 
@@ -187,7 +187,7 @@ final class ChannelUpdater_Tests: XCTestCase {
         try database.createChannel(cid: cid, withMessages: false)
         // Local only message
         try database.createMessage(cid: cid, localState: .sendingFailed)
-        try database.createMessage(cid: cid, localState: .deletingFailed)
+        try database.createMessage(cid: cid, localState: .pendingSend)
         // Not local only message
         try database.createMessage(cid: cid, localState: .syncing)