+++ Tests/StreamChatUITests/Mocks/ChatMessageList/ChatMessageListVCDataSource_Mock.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2023 Stream.io Inc. All rights reserved.
+// Copyright © 2024 Stream.io Inc. All rights reserved.
 //
 
 @testable import StreamChat
@@ -29,7 +29,7 @@ class ChatMessageListVCDataSource_Mock: ChatMessageListVCDataSource {
     }
 
     func chatMessageListVC(_ vc: StreamChatUI.ChatMessageListVC, messageAt indexPath: IndexPath) -> StreamChat.ChatMessage? {
-        messages[indexPath.item]
+        messages[safe: indexPath.item]
     }
 
     func chatMessageListVC(_ vc: StreamChatUI.ChatMessageListVC, messageLayoutOptionsAt indexPath: IndexPath) -> StreamChatUI.ChatMessageLayoutOptions {