+++ Tests/StreamChatUITests/SnapshotTests/ChatMessageList/Attachments/ChatFileAttachmentListView+ItemView_Tests.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2023 Stream.io Inc. All rights reserved.
+// Copyright © 2024 Stream.io Inc. All rights reserved.
 //
 
 import StreamChat
@@ -26,6 +26,20 @@ final class ChatFileAttachmentListViewItemView_Tests: XCTestCase {
 
     func test_appearance_pdf() {
         fileAttachmentView.content = .mock(id: .unique)
+        AssertSnapshot(fileAttachmentView, variants: [.defaultLight])
+    }
+
+    func test_appearance_pdf_whenUploadingStateIsNil() {
+        fileAttachmentView.content = .mock(id: .unique, localState: nil)
+        AssertSnapshot(fileAttachmentView, variants: [.defaultLight])
+    }
+
+    func test_appearance_pdf_whenSizeIsZero() {
+        fileAttachmentView.content = .mock(
+            id: .unique,
+            file: AttachmentFile(type: .pdf, size: 0, mimeType: "application/pdf"),
+            localState: nil
+        )
         AssertSnapshot(fileAttachmentView, variants: [.defaultLight])
     }