+++ Tests/StreamChatUITests/SnapshotTests/ChatMessageList/ChatMessageListVC_Tests.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2023 Stream.io Inc. All rights reserved.
+// Copyright © 2024 Stream.io Inc. All rights reserved.
 //
 
 @testable import StreamChat
@@ -42,6 +42,7 @@ final class ChatMessageListVC_Tests: XCTestCase {
     override func tearDown() {
         mockedDataSource = nil
         mockedDelegate = nil
+        AttachmentViewCatalog_Mock.attachmentViewInjectorClassForCallCount = 0
         super.tearDown()
     }
 
@@ -475,7 +476,21 @@ final class ChatMessageListVC_Tests: XCTestCase {
         XCTAssertEqual(mockedListView.scrollToBottomCallCount, 1)
     }
 
-    func test_updateMessages_whenNewestMessageMovedByCurrentUser_whenFistPageNotLoaded_shouldScrollToBottom() {
+    func test_updateMessages_whenNewestMessageInsertedByCurrentUser_whenMultipleInsertions_shouldScrollToBottom() {
+        mockedDataSource.mockedIsFirstPageLoaded = true
+        mockedListView.mockIsLastCellFullyVisible = true
+
+        sut.updateMessages(with: [
+            .insert(.mock(isSentByCurrentUser: false), index: .init(item: 0, section: 0)),
+            .insert(.mock(isSentByCurrentUser: false), index: .init(item: 0, section: 0)),
+            .insert(.mock(isSentByCurrentUser: true), index: .init(item: 0, section: 0))
+        ])
+        mockedListView.updateMessagesCompletion?()
+
+        XCTAssertEqual(mockedListView.scrollToBottomCallCount, 0)
+    }
+
+    func test_updateMessages_whenNewestMessageMovedByCurrentUser_whenFistPageNotLoaded_shouldNotScrollToBottom() {
         mockedDataSource.mockedIsFirstPageLoaded = false
         mockedListView.mockIsLastCellFullyVisible = true
 
@@ -1081,5 +1096,61 @@ final class ChatMessageListVC_Tests: XCTestCase {
         // Then
         sut.handlePan(.init())
         XCTAssertEqual(handlerMock.handleCallCount, 0)
+    }
+
+    // MARK: - attachmentViewInjectorClassForMessage
+
+    func test_attachmentViewInjectorClassForMessage_shouldAskForAttachmentInjector() {
+        sut.components.attachmentViewCatalog = AttachmentViewCatalog_Mock.self
+        mockedDataSource.messages = [.unique]
+
+        _ = sut.attachmentViewInjectorClassForMessage(at: .init(item: 0, section: 0))
+
+        XCTAssertEqual(AttachmentViewCatalog_Mock.attachmentViewInjectorClassForCallCount, 1)
+    }
+
+    func test_attachmentViewInjectorClassForMessage_whenMessageIsNil_returnsNil() {
+        sut.components.attachmentViewCatalog = AttachmentViewCatalog_Mock.self
+        mockedDataSource.messages = []
+
+        _ = sut.attachmentViewInjectorClassForMessage(at: .init(item: 0, section: 0))
+
+        XCTAssertEqual(AttachmentViewCatalog_Mock.attachmentViewInjectorClassForCallCount, 0)
+    }
+
+    func test_attachmentViewInjectorClassForMessage_whenMessageIsDeleted_returnsNil() {
+        sut.components.attachmentViewCatalog = AttachmentViewCatalog_Mock.self
+        mockedDataSource.messages = [.mock(deletedAt: .unique)]
+
+        _ = sut.attachmentViewInjectorClassForMessage(at: .init(item: 0, section: 0))
+
+        XCTAssertEqual(AttachmentViewCatalog_Mock.attachmentViewInjectorClassForCallCount, 0)
+    }
+
+    func test_attachmentViewInjectorClassForMessage_whenMessageIsSystem_returnsNil() {
+        sut.components.attachmentViewCatalog = AttachmentViewCatalog_Mock.self
+        mockedDataSource.messages = [.mock(type: .system)]
+
+        _ = sut.attachmentViewInjectorClassForMessage(at: .init(item: 0, section: 0))
+
+        XCTAssertEqual(AttachmentViewCatalog_Mock.attachmentViewInjectorClassForCallCount, 0)
+    }
+
+    func test_attachmentViewInjectorClassForMessage_whenMessageIsError_returnsNil() {
+        sut.components.attachmentViewCatalog = AttachmentViewCatalog_Mock.self
+        mockedDataSource.messages = [.mock(type: .error, isBounced: false)]
+
+        _ = sut.attachmentViewInjectorClassForMessage(at: .init(item: 0, section: 0))
+
+        XCTAssertEqual(AttachmentViewCatalog_Mock.attachmentViewInjectorClassForCallCount, 0)
+    }
+}
+
+class AttachmentViewCatalog_Mock: AttachmentViewCatalog {
+    static var mockedInjector: AttachmentViewInjector.Type?
+    static var attachmentViewInjectorClassForCallCount = 0
+    override class func attachmentViewInjectorClassFor(message: ChatMessage, components: Components) -> AttachmentViewInjector.Type? {
+        attachmentViewInjectorClassForCallCount += 1
+        return mockedInjector
     }
 }